+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 1. mctpd: Add support for endpoint recovery
 2. mctpd: Allow recovery of devices reporting a nil UUID for development
+3. mctpd: Allow configuring .Connectivity as writable for development
 
 ### Changed